requirements.txt
# Core framework
fastapi==0.127.1
uvicorn[standard]>=0.32.0
flask>=3.0.0

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication & Security
PyJWT>=2.10.1
bcrypt>=4.0.1

# Rate limiting
redis>=5.0.1
slowapi>=0.1.9

# Utils
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings>=2.7.1
python-dotenv>=1.0.0
jinja2>=3.1.0
openpyxl>=3.1.0
